Split out AppBarProfileOverflowMenu
